+++ source/Messaging.Application/Messaging.Application.csproj
@@ -48,16 +48,28 @@ limitations under the License.
     </ItemGroup>
 
     <ItemGroup>
-      <EmbeddedResource Include="SchemaStore\Schemas\Xml\urn-ediel-org-structure-confirmrequestchangeofsupplier-0-1.xsd" />
-      <EmbeddedResource Include="SchemaStore\Schemas\Xml\urn-ediel-org-structure-genericnotification-0-1.xsd" />
-      <EmbeddedResource Include="SchemaStore\Schemas\Xml\urn-ediel-org-structure-rejectrequestchangeofsupplier-0-1.xsd" />
-      <EmbeddedResource Include="SchemaStore\Schemas\Xml\urn-ediel-org-structure-requestchangeofsupplier-0-1.xsd" />
-      <EmbeddedResource Include="SchemaStore\Schemas\Xml\urn-entsoe-eu-local-extension-types.xsd" />
-      <EmbeddedResource Include="SchemaStore\Schemas\Xml\urn-entsoe-eu-wgedi-codelists.xsd" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <Folder Include="SchemaStore\Schemas\Json" />
+      <EmbeddedResource Include="SchemaStore\Schemas\Xml\urn-ediel-org-structure-confirmrequestchangeofsupplier-0-1.xsd">
+        <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+      </EmbeddedResource>
+      <EmbeddedResource Include="SchemaStore\Schemas\Xml\urn-ediel-org-structure-genericnotification-0-1.xsd">
+        <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+      </EmbeddedResource>
+      <EmbeddedResource Include="SchemaStore\Schemas\Xml\urn-ediel-org-structure-rejectrequestchangeofsupplier-0-1.xsd">
+        <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+      </EmbeddedResource>
+      <EmbeddedResource Include="SchemaStore\Schemas\Xml\urn-ediel-org-structure-requestchangeofsupplier-0-1.xsd">
+        <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+      </EmbeddedResource>
+      <EmbeddedResource Include="SchemaStore\Schemas\Xml\urn-entsoe-eu-local-extension-types.xsd">
+        <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+      </EmbeddedResource>
+      <EmbeddedResource Include="SchemaStore\Schemas\Xml\urn-entsoe-eu-wgedi-codelists.xsd">
+        <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+      </EmbeddedResource>
+      <None Remove="SchemaStore\Schemas\Json\Request-Change-of-Supplier-assembly-model.schema.json" />
+      <EmbeddedResource Include="SchemaStore\Schemas\Json\Request-Change-of-Supplier-assembly-model.schema.json">
+        <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+      </EmbeddedResource>
     </ItemGroup>
 
 </Project>
